pyproject.toml
[tool.poetry]
name = "DocDig"
version = "0.1.0"
description = ""
authors = ["Benedikt Veith <bv@scherer-software.de>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

pathlib = "^1.0.1"
pdf2image = "^1.17.0"
tensorflow = "^2.17.0"
ipython = "^8.28.0"
opencv-python = "^4.10.0.84"
python-dotenv = "^1.0.1"
pandas = "^2.2.3" 
numpy = "^1.24.2"

torch = "^2.3.0"
torchvision = "^0.18.0"
scikit-learn = "^1.2.1"
tqdm = "^4.65.0"
Cython = "^0.29.33"
matplotlib = "^3.7.0"
pycocotools = "^2.0.6"
Pillow = "^9.4.0"
editdistance = "^0.6.2"
scikit-image = "^0.20.0"
scipy = "^1.10.1"
PyMuPDF = "^1.21.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
